Fixing an issue with slow gallery reset
When switching to another folder, previous images
will stay for a while. So it looks like the page changes,
but contents changes after a sec or so.
The reason is that we filter empty image arrays from
image-list datasource observable, which means that
when we clear the store, the component doesn't receive
an update